+++ x/oracle/keeper/aggregate_test.go
@@ -182,7 +182,7 @@ func (s *KeeperTestSuite) TestSetAggregate() {
 	err = s.oracleKeeper.SetAggregate(s.ctx, report)
 	s.NoError(err)
 
-	res, err := s.oracleKeeper.Aggregates.Get(s.ctx, collections.Join(queryId, timestamp.UnixMilli()))
+	res, err := s.oracleKeeper.Aggregates.Get(s.ctx, collections.Join(queryId, uint64(timestamp.UnixMilli())))
 	s.NoError(err)
 	s.Equal(encodeValue(96.50), res.AggregateValue)
 	s.Equal(int64(100000000), res.ReporterPower)
@@ -279,7 +279,7 @@ func (s *KeeperTestSuite) TestGetTimestampBefore() {
 			for _, v := range tc.timestamps {
 				err := s.oracleKeeper.Aggregates.Set(
 					s.ctx,
-					collections.Join(queryId, v.UnixMilli()),
+					collections.Join(queryId, uint64(v.UnixMilli())),
 					types.Aggregate{},
 				)
 				s.Require().NoError(err)
@@ -357,7 +357,7 @@ func (s *KeeperTestSuite) TestGetTimestampAfter() {
 			for _, v := range tc.timestamps {
 				err := s.oracleKeeper.Aggregates.Set(
 					s.ctx,
-					collections.Join(queryId, v.UnixMilli()),
+					collections.Join(queryId, uint64(v.UnixMilli())),
 					types.Aggregate{},
 				)
 				s.Require().NoError(err)
@@ -384,7 +384,7 @@ func (s *KeeperTestSuite) TestGetAggregatedReportsByHeight() {
 	s.NoError(err)
 
 	s.ctx = s.ctx.WithBlockHeight(15)
-	aggregates := s.oracleKeeper.GetAggregatedReportsByHeight(s.ctx, int64(10))
+	aggregates := s.oracleKeeper.GetAggregatedReportsByHeight(s.ctx, uint64(10))
 	s.NotEqual(0, len(aggregates))
 	s.Equal(*aggregate, aggregates[0])
 }
