+++ x/oracle/keeper/msg_server_submit_value_test.go
@@ -69,7 +69,7 @@ func (s *KeeperTestSuite) TestSubmitValue() (sdk.AccAddress, []byte) {
 		Value:           value,
 		Timestamp:       s.ctx.BlockTime(),
 		Cyclelist:       true,
-		BlockNumber:     s.ctx.BlockHeight(),
+		BlockNumber:     uint64(s.ctx.BlockHeight()),
 	}
 	expectedReport := types.QueryMicroReportsResponse{
 		MicroReports: []types.MicroReport{microReport},
@@ -282,7 +282,7 @@ func (s *KeeperTestSuite) TestDirectReveal() {
 		QueryType:             "SpotPrice",
 	}
 	reporter := sample.AccAddressBytes()
-	votingPower := sdk.TokensToConsensusPower(math.NewInt(1_000_000), layertypes.PowerReduction)
+	votingPower := uint64(sdk.TokensToConsensusPower(math.NewInt(1_000_000), layertypes.PowerReduction))
 	isBridgeDeposit := true
 	ctx = ctx.WithBlockTime(ctx.BlockTime().Add(time.Hour))
 	err = k.DirectReveal(ctx, query, qDataBz, value, reporter, votingPower, !isBridgeDeposit)
