+++ x/reporter/keeper/distribution_test.go
@@ -18,14 +18,14 @@ import (
 
 func TestDivvyingTips(t *testing.T) {
 	k, _, _, _, ctx, _ := setupKeeper(t)
-	height := int64(10)
+	height := uint64(10)
 	val1Address := sample.AccAddressBytes()
 	vals := simtestutil.ConvertAddrsToValAddrs([]sdk.AccAddress{val1Address})
 	val1 := vals[0]
 	addr := sample.AccAddressBytes()
 	addr2 := sample.AccAddressBytes()
 	reporter1 := types.NewReporter(math.LegacyZeroDec(), math.OneInt())
-	ctx = ctx.WithBlockHeight(height)
+	ctx = ctx.WithBlockHeight(int64(height))
 
 	err := k.Reporters.Set(ctx, addr, reporter1)
 	require.NoError(t, err)
